misc/building_ai_applications/demo/pyproject.toml
[tool.poetry]
name = "generating-guide-demo"
version = "0.0.1"
description = ""
authors = ["eddiechayes <eddie@konfigthis.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
jupyter = "^1.0.0"
langchain = "^0.1.16"
langchain-community = "^0.0.32"
langchainhub = "^0.1.15"
langchain-openai = "^0.1.2"
chromadb = "^0.4.24"
bs4 = "^0.0.2"
jupyterlab = "^4.1.6"
langchain-text-splitters = "^0.0.1"
langchain-chroma = "^0.1.0"
jsonref = "^1.1.0"
langchain-pinecone = "^0.1.0"
lark = "^1.1.9"
pinecone-client = "^3.2.2"
arize-phoenix = "^3.22.0"
langgraph = "^0.0.39"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
